rename repository to store
