+++ internal/store/postgres.go
@@ -1,4 +1,4 @@
-package storage
+package store
 
 import (
 	"context"
